requirements.txt
bs4
loguru
arsenic
asyncio
aiohttp
structlog
python-dotenv
fake_useragent
webdriver_manager